- Rename provider folder to civo - Change .travis.yml to use Go 1.13
