+++ civo/resource_network.go
@@ -1,4 +1,4 @@
-package provider
+package civo
 
 import (
 	"fmt"
