+++ go.mod
@@ -2,7 +2,7 @@ module github.com/civo/terraform-provider-civo
 
 require (
 	github.com/aws/aws-sdk-go v1.25.39
-	github.com/civo/civogo v0.0.0-20200123124629-f66cbcbb6f2d
+	github.com/civo/civogo v0.0.0-20200123135111-b3aba767c3d7
 	github.com/hashicorp/terraform-plugin-sdk v1.3.0
 	golang.org/x/crypto v0.0.0-20200115085410-6d4e4cb37c7d // indirect
 )
