+++ main.go
@@ -1,7 +1,7 @@
 package main
 
 import (
-	"github.com/civo/terraform-provider-civo/provider"
+	"github.com/civo/terraform-provider-civo/civo"
 	"github.com/hashicorp/terraform-plugin-sdk/plugin"
 	"github.com/hashicorp/terraform-plugin-sdk/terraform"
 )
@@ -9,7 +9,7 @@ import (
 func main() {
 	plugin.Serve(&plugin.ServeOpts{
 		ProviderFunc: func() terraform.ResourceProvider {
-			return provider.Provider()
+			return civo.Provider()
 		},
 	})
 }
